Update method calculate payment by salary and working days to return BigDecimal

+++ src/main/java/rdguez/eduardo/upax/dto/EmployeePaymentDto.java
@@ -2,9 +2,11 @@ package rdguez.eduardo.upax.dto;
 
 import rdguez.eduardo.upax.model.response.EmployeePaymentResponse;
 
+import java.math.BigDecimal;
+
 public class EmployeePaymentDto {
 
-  public static EmployeePaymentResponse toResponse(double payment) {
+  public static EmployeePaymentResponse toResponse(BigDecimal payment) {
     return EmployeePaymentResponse.builder()
       .payment(payment)
       .success(true)
